requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python==4.8.1.78
ultralytics==8.0.196
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
numpy<2.0.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
boto3>=1.34.0
botocore>=1.34.0
